# [flutter_tools] improve hash performance