--- a/packages/flutter_tools/lib/src/build_system/build_system.dart
+++ b/packages/flutter_tools/lib/src/build_system/build_system.dart
@@ -802,10 +802,12 @@ class _BuildInstance {
       // If we were missing the depfile, resolve input files after executing the
       // target so that all file hashes are up to date on the next run.
       if (node.missingDepfile) {
+        await null; // TODO(jonahwilliams): update expectations in resident_runner_test.dart
         fileCache.diffFileList(node.inputs);
       }
 
       // Always update hashes for output files.
+      await null; // TODO(jonahwilliams): update expectations in resident_runner_test.dart
       fileCache.diffFileList(node.outputs);
       node.target._writeStamp(node.inputs, node.outputs, environment);
       updateGraph();
@@ -1075,6 +1077,7 @@ class Node {
     // If we have files to diff, compute them asynchronously and then
     // update the result.
     if (sourcesToDiff.isNotEmpty) {
+      await null; // TODO(jonahwilliams): update expectations in resident_runner_test.dart
       final List<File> dirty = fileStore.diffFileList(sourcesToDiff);
       if (dirty.isNotEmpty) {
         invalidatedReasons.add(InvalidatedReason.inputChanged);