## Description


new hash performance (fully cached)
```
[  +10 ms]   [  +10 ms] Done initializing file store
[  +33 ms]   [  +33 ms] Skipping target: gen_localizations
[ +547 ms]   [ +546 ms] Skipping target: unpack_windows
[        ]   [        ] Skipping target: kernel_snapshot
[   +6 ms]   [   +7 ms] Skipping target: aot_elf_release
[   +4 ms]   [   +4 ms] Skipping target: windows_aot_bundle
[   +9 ms]   [   +9 ms] Skipping target: release_bundle_windows_assets

 Measure-Command { flutter build windows }

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 4
Milliseconds      : 764
Ticks             : 47645191
TotalDays         : 5.51448969907407E-05
TotalHours        : 0.00132347752777778
TotalMinutes      : 0.0794086516666667
TotalSeconds      : 4.7645191
TotalMilliseconds : 4764.5191
```



Current hash performance (fully cached)
```
[  +39 ms]   [  +42 ms] Initializing file store
[  +12 ms]   [  +12 ms] Done initializing file store
[  +33 ms]   [  +33 ms] Skipping target: gen_localizations
[ +892 ms]   [ +892 ms] Skipping target: kernel_snapshot
[  +71 ms]   [  +71 ms] Skipping target: aot_elf_release
[  +69 ms]   [  +69 ms] Skipping target: windows_aot_bundle
[+1509 ms]   [+1509 ms] Skipping target: unpack_windows
[  +28 ms]   [  +28 ms] Skipping target: release_bundle_windows_assets
[        ]   [        ] Persisting file store

Measure-Command { flutter build windows }

Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 6
Milliseconds      : 697
Ticks             : 66974585
TotalDays         : 7.7516880787037E-05
TotalHours        : 0.00186040513888889
TotalMinutes      : 0.111624308333333
TotalSeconds      : 6.6974585
TotalMilliseconds : 6697.4585

```