@jonahwilliams 2020-11-10
Made this actually implement md5 correctly