@jonahwilliams 2020-11-11
I tried a few more micro-optimizations, but I couldn't make any more progress. This is still 30-40% faster than package:crypto so I'd like to land this and move on